requirements.txt
asyncio
datetime
playsound
discord.py-self
python3-anticaptcha